Owners feature

See merge request dokerplp/booking!13

+++ src/main/java/com/example/booking/service/HotelService.java
@@ -11,14 +11,17 @@ import com.example.booking.exception.HotelNotFoundException;
 import com.example.booking.mapper.HotelReadMapper;
 import com.example.booking.repository.HotelRepository;
 import com.example.booking.repository.UserRepository;
-import lombok.AllArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.nio.file.AccessDeniedException;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 
@@ -33,8 +36,8 @@ public class HotelService {
     private final HotelReadMapper hotelReadMapper;
 
     public Optional<HotelInfoDto> findHotelInfo(long id) {
-       return hotelRepo.findHotelById(id)
-               .map(hotelReadMapper::map);
+        return hotelRepo.findHotelById(id)
+                .map(hotelReadMapper::map);
     }
 
     public List<HotelInfoDto> findHotels(int page) {
@@ -70,38 +73,51 @@ public class HotelService {
 
     @Transactional
     public Long saveHotel(HotelCreateDto hotelCreateDto) {
-        Optional<User> optionalOwner = userRepository.findById(hotelCreateDto.getOwnerId());
+        UserDetails principal = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Optional<User> optionalOwner = userRepository.findByLogin(principal.getUsername());
         User owner = optionalOwner.get();
 
-        Hotel hotel = new Hotel();
-        hotel.setName(hotelCreateDto.getName());
-        hotel.setCity(hotelCreateDto.getCity());
-        hotel.setHotelClass(hotelCreateDto.getHotelClass());
-        hotel.setOwner(owner);
+        Hotel hotel = Hotel.builder()
+                .name(hotelCreateDto.getName())
+                .city(hotelCreateDto.getCity())
+                .hotelClass(hotelCreateDto.getHotelClass())
+                .owner(owner)
+                .build();
 
         Hotel savedHotel = hotelRepo.save(hotel);
         return savedHotel.getId();
     }
 
     @Transactional
-    public void updateHotel(HotelUpdateDto hotelDto, long hotelId) {
+    public void updateHotel(HotelUpdateDto hotelDto, long hotelId) throws AccessDeniedException {
         Optional<Hotel> optionalHotel = hotelRepo.findById(hotelId);
         if (optionalHotel.isEmpty()) {
             throw new HotelNotFoundException(hotelId);
         }
         Hotel hotel = optionalHotel.get();
+        UserDetails principal = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 
-        hotel.setName(hotelDto.getName());
-        hotel.setCity(hotelDto.getCity());
-        hotel.setHotelClass(hotelDto.getHotelClass());
+        if (Objects.equals(hotel.getOwner().getLogin(), principal.getUsername())) {
+            hotel.setName(hotelDto.getName());
+            hotel.setCity(hotelDto.getCity());
+            hotel.setHotelClass(hotelDto.getHotelClass());
 
-        hotelRepo.save(hotel);
+            hotelRepo.save(hotel);
+        } else {
+            throw new AccessDeniedException("You do not have permission to modify this hotel");
+        }
     }
 
     @Transactional
-    public void deleteHotel(long hotelId) {
+    public void deleteHotel(long hotelId) throws AccessDeniedException {
         if (hotelRepo.existsById(hotelId)) {
-            hotelRepo.deleteById(hotelId);
+            UserDetails principal = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+            Hotel hotel = hotelRepo.findHotelById(hotelId).get();
+            if (Objects.equals(hotel.getOwner().getLogin(), principal.getUsername())) {
+                hotelRepo.deleteById(hotelId);
+            } else {
+                throw new AccessDeniedException("You do not have permission to delete this hotel");
+            }
         } else {
             throw new HotelNotFoundException(hotelId);
         }
